Add invoice PDF download to dashboard orders page
- Extract invoice PDF generation into shared @/utils/invoice.js utility
  (generateInvoicePDF + loadJsPDF) so both COD page and dashboard can use it
- Wire "Download invoice" button in dashboard orders with loading spinner,
  disabled state while generating, and error handling
- COD order page: replaced ~350 lines of inline jsPDF code with 10-line
  wrapper calling shared utility
- jsPDF CDN loaded lazily on mount via loadJsPDF() helper
- Each order card shows Download icon + "Preparing..." spinner state
